requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
crispy-forms-gds==0.2.2
Django==3.2.4
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-utils-six==2.0
idna==2.10
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
validate-email==1.3
